Chore: add resetForm on Cancel button and move tabs
