+++ frontend/src/pages/Administration/Settings/components/tabs/ConfigurationTab.tsx
@@ -23,8 +23,8 @@ import { Check, ClipboardCopy } from 'lucide-react';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import type { SettingsTabProps } from '../SettingsConfig';
-import SettingsSection from '../components/SettingsSection';
+import type { SettingsTabProps } from '../../SettingsConfig';
+import SettingsSection from '../SettingsSection';
 
 import SelectDropdown from '@/components/ui/forms/SelectDropdown';
 import TextInput from '@/components/ui/forms/TextInput';
