+++ frontend/src/pages/Administration/Settings/components/tabs/DisplaysTab.tsx
@@ -21,8 +21,8 @@
 
 import { useTranslation } from 'react-i18next';
 
-import type { SettingsTabProps } from '../SettingsConfig';
-import SettingsSection from '../components/SettingsSection';
+import type { SettingsTabProps } from '../../SettingsConfig';
+import SettingsSection from '../SettingsSection';
 
 import NumberInput from '@/components/ui/forms/NumberInput';
 import SelectDropdown from '@/components/ui/forms/SelectDropdown';
