+++ frontend/src/pages/Administration/Settings/components/tabs/TroubleshootingTab.tsx
@@ -21,8 +21,8 @@
 
 import { useTranslation } from 'react-i18next';
 
-import type { SettingsTabProps } from '../SettingsConfig';
-import SettingsSection from '../components/SettingsSection';
+import type { SettingsTabProps } from '../../SettingsConfig';
+import SettingsSection from '../SettingsSection';
 
 import DatePickerInput from '@/components/ui/forms/DatePickerInput';
 import SelectDropdown from '@/components/ui/forms/SelectDropdown';
